mainPage->chat
메인인페이지에서 list 누를 시 popup으로 넘어가는 거 하려는데 안됩니다.. ㅜㅜ
